# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nyx LogBook System API
  version: 1.0.0
  description: Enriched OpenAPI 3.1 specification generated from actual route implementations.
servers:
- url: /
paths:
...
  /admin/competencies:
    get:
      summary: GET /admin/competencies
      description: Endpoint implemented in src\app\api\admin\competencies\route.ts.
      tags:
      - Admin
      responses:
        '401':
          description: Error response
          content:
            application/json:
              schema: &id017
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '409':
          description: Error response
          content:
            application/json:
              schema: &id018
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles: &id019
      - ADMIN
    post:
      summary: POST /admin/competencies
      description: Endpoint implemented in src\app\api\admin\competencies\route.ts.
      tags:
      - Admin
      responses:
        '401':
          description: Error response
          content:
            application/json:
              schema: *id017
        '409':
          description: Error response
          content:
            application/json:
              schema: *id018
      security:
      - nextAuthSession: []
      x-roles: *id019
  /admin/competency-assessments/{id}:
    put:
      summary: PUT /admin/competency-assessments/{id}
      description: Endpoint implemented in src\app\api\admin\competency-assessments\[id]\route.ts.
      tags:
      - Admin
      responses:
        '401':
          description: Error response
          content:
            application/json:
              schema: &id020
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '403':
          description: Error response
          content:
            application/json:
              schema: &id021
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '500':
          description: Error response
          content:
            application/json:
              schema: &id022
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles: &id023
      - ADMIN
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Path parameter id.
    delete:
      summary: DELETE /admin/competency-assessments/{id}
      description: Endpoint implemented in src\app\api\admin\competency-assessments\[id]\route.ts.
      tags:
      - Admin
      responses:
        '401':
          description: Error response
          content:
            application/json:
              schema: *id020
        '403':
          description: Error response
          content:
            application/json:
              schema: *id021
        '500':
          description: Error response
          content:
            application/json:
              schema: *id022
      security:
      - nextAuthSession: []
      x-roles: *id023
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Path parameter id.
  /admin/competency-assessments:
    get:
      summary: GET /admin/competency-assessments
      description: Endpoint implemented in src\app\api\admin\competency-assessments\route.ts.
      tags:
      - Admin
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: &id024
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '201':
          description: Error response
          content:
            application/json:
              schema: &id025
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '401':
          description: Error response
          content:
            application/json:
              schema: &id026
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles: &id027
      - ADMIN
    post:
      summary: POST /admin/competency-assessments
      description: Endpoint implemented in src\app\api\admin\competency-assessments\route.ts.
      tags:
      - Admin
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: *id024
        '201':
          description: Error response
          content:
            application/json:
              schema: *id025
        '401':
          description: Error response
          content:
            application/json:
              schema: *id026
      security:
      - nextAuthSession: []
      x-roles: *id027
...
  /admin/curriculum/import:
    post:
      summary: POST /admin/curriculum/import
      description: Endpoint implemented in src\app\api\admin\curriculum\import\route.ts.
      tags:
      - Admin
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles:
      - ADMIN
...
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie used for authenticated requests.